requirements.txt
amqp==2.3.2
aniso8601==3.0.2
asn1crypto==0.24.0
billiard==3.5.0.5
celery==4.2.1
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
cookies==2.2.1
cryptography==2.3.1
defusedxml==0.5.0
Django==2.1.3
django-admin-shortcuts==2.0.0
django-allauth==0.37.1
django-background-tasks==1.2.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap3==11.0.0
django-compat==1.0.15
django-crispy-forms==1.7.2
django-datetimepicker==3.14
django-extensions==2.1.6
django-filter==2.0.0
django-modeladmin-reorder==0.3.1
django-report-builder==6.2.2
django-rest-auth==0.9.3
django-tables2==2.0.2
django-widget-tweaks==1.4.3
djangocms-admin-style==1.2.9
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drfdocs==0.0.6
et-xmlfile==1.0.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
httpie==0.9.9
humanize==0.5.1
idna==2.7
jdcal==1.4
kombu==4.2.2.post1
Markdown==2.6.9
mysqlclient==1.3.13
numpy==1.15.4
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.4.9
pandas==0.23.4
Pillow==5.4.1
pipenv==2018.7.1
postgres==2.2.2
promise==2.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.2
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
redis==3.0.1
requests==2.18.4
requests-oauthlib==1.0.0
responses==0.9.0
rollbar==0.14.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
tablib==0.12.1
typing==3.6.6
unicodecsv==0.14.1
urllib3==1.22
vine==1.2.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
xlrd==1.1.0
xlwt==1.3.0

